경북대 BE_주보경 2주차 과제 (1단계) (#211)
* Feat: step0 코드 준비

* Feat: API restful하게 수정

* Docs: README.md 작성

* Chore: valid를 위한 라이브러리 추가

* Feat: 유효성 검사 로직 추가

* Feat: 유효성 검사 로직 메시지 추가

* Feat: 이름 유효성 체크 로직 추가

* Feat: 컨트롤러 에러 핸들링

* Feat: 에러 핸들링 기능 구현

* Feat: id,name 에러 체크 로직

+++ src/main/java/gift/domain/product/ProductJdbcRepository.java
@@ -5,7 +5,8 @@ import org.springframework.context.annotation.Primary;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
 
-import java.util.ArrayList;
+import java.util.NoSuchElementException;
+
 
 @Repository
 @Primary
@@ -36,6 +37,9 @@ public class ProductJdbcRepository implements ProductRepository {
 
     @Override
     public Product findById(Long id) {
+        if (isNotValidProductId(id)){
+            throw new NoSuchElementException("유효하지 않은 id입니다.");
+        }
         String sql = "SELECT * FROM product where id = ?";
         return jdbcTemplate.queryForObject(sql, (rs, rowNum) -> {
             String name = rs.getString("name");
@@ -47,17 +51,28 @@ public class ProductJdbcRepository implements ProductRepository {
 
     @Override
     public void deleteById(Long id) {
+        if (isNotValidProductId(id)){
+            throw new NoSuchElementException("유효하지 않은 id입니다.");
+        }
         String sql = "DELETE FROM product where id = ?";
         jdbcTemplate.update(sql, id);
     }
 
     @Override
     public int update(Long id, ProductRequestDto requestDto) {
+        if (isNotValidProductId(id)){
+            throw new NoSuchElementException("유효하지 않은 id입니다.");
+        }
         String sql = "UPDATE product set name = ?, price = ?, img_url = ? where id = ?";
         return jdbcTemplate.update(sql,
                 requestDto.getName(),
                 requestDto.getPrice(),
                 requestDto.getImgUrl(),
                 id);
+    }
+  
+    private boolean isNotValidProductId(Long id){
+        String sql = "SELECT * FROM product WHERE id=?";
+        return jdbcTemplate.query(sql, (rs,rowNum)-> 0, id).isEmpty();
     }
 }
